+++ .idea/workspace.xml
@@ -4,7 +4,18 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="038c1280-d02a-4ac6-bcd1-233dc77f04ba" name="Changes" comment="" />
+    <list default="true" id="038c1280-d02a-4ac6-bcd1-233dc77f04ba" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/src/produit/clazz/ProduitToxique.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/produit/interfaze/Toxique.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/Main.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/clazz/Produit.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/produit/clazz/Produit.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/clazz/ProduitFrais.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/produit/clazz/ProduitFrais.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/clazz/ProduitFraisNonToxic.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/produit/clazz/ProduitFraisNonToxique.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/clazz/ProduitFraisToxic.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/produit/clazz/ProduitFraisToxique.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/clazz/ProduitNonToxic.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/produit/clazz/ProduitNonToxique.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/clazz/ProduitToxic.java" beforeDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -14,6 +25,7 @@
     <option name="RECENT_TEMPLATES">
       <list>
         <option value="Class" />
+        <option value="Interface" />
       </list>
     </option>
   </component>
@@ -92,7 +104,14 @@
       <option name="project" value="LOCAL" />
       <updated>1669985107127</updated>
     </task>
-    <option name="localTasksCounter" value="7" />
+    <task id="LOCAL-00007" summary="Modif Package Organisation">
+      <created>1669985421040</created>
+      <option name="number" value="00007" />
+      <option name="presentableId" value="LOCAL-00007" />
+      <option name="project" value="LOCAL" />
+      <updated>1669985421040</updated>
+    </task>
+    <option name="localTasksCounter" value="8" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -112,6 +131,7 @@
     <MESSAGE value="Adding diff empty files" />
     <MESSAGE value="Adding Products" />
     <MESSAGE value="Adding Branch" />
-    <option name="LAST_COMMIT_MESSAGE" value="Adding Branch" />
+    <MESSAGE value="Modif Package Organisation" />
+    <option name="LAST_COMMIT_MESSAGE" value="Modif Package Organisation" />
   </component>
 </project>
