--- src/clazz/ProduitFraisNonToxic.java
@@ -1,7 +1,0 @@
-package clazz;
-
-public class ProduitFraisNonToxic extends ProduitFrais {
-    public ProduitFraisNonToxic(String nom, int prix, String marque, int temperature) {
-        super(nom, prix, marque, temperature);
-    }
-}
