--- src/clazz/ProduitFraisToxic.java
@@ -1,7 +1,0 @@
-package clazz;
-
-public class ProduitFraisToxic extends ProduitFrais {
-    public ProduitFraisToxic(String nom, int prix, String marque, int temperature) {
-        super(nom, prix, marque, temperature);
-    }
-}
