--- src/clazz/ProduitNonToxic.java
@@ -1,7 +1,0 @@
-package clazz;
-
-public class ProduitNonToxic extends Produit {
-    public ProduitNonToxic(String nom, int prix, String marque) {
-        super(nom, prix, marque);
-    }
-}
