--- src/clazz/ProduitToxic.java
@@ -1,7 +1,0 @@
-package clazz;
-
-public class ProduitToxic extends Produit {
-    public ProduitToxic(String nom, int prix, String marque) {
-        super(nom, prix, marque);
-    }
-}
