+++ src/produit/clazz/ProduitFrais.java
@@ -1,0 +1,21 @@
+package produit.clazz;
+
+public abstract class ProduitFrais extends Produit {
+    private int temperature;
+
+    //Constructeur
+    public ProduitFrais(String nom, int prix, String marque, int temperature) {
+        super(nom, prix, marque);
+        this.temperature = temperature;
+    }
+
+    //<========================= Getter / Setter =========================>
+
+    public int getTemperature() {
+        return temperature;
+    }
+
+    public void setTemperature(int temperature) {
+        this.temperature = temperature;
+    }
+}
