+++ src/produit/clazz/ProduitFraisNonToxique.java
@@ -1,0 +1,11 @@
+package produit.clazz;
+
+public class ProduitFraisNonToxique extends ProduitFrais {
+
+    //Constructeur
+    public ProduitFraisNonToxique(String nom, int prix, String marque, int temperature) {
+        super(nom, prix, marque, temperature);
+    }
+
+
+}
