+++ src/produit/clazz/ProduitFraisToxique.java
@@ -1,0 +1,10 @@
+package produit.clazz;
+
+public class ProduitFraisToxique extends ProduitFrais {
+
+    //Constructeur
+    public ProduitFraisToxique(String nom, int prix, String marque, int temperature) {
+        super(nom, prix, marque, temperature);
+    }
+
+}
