+++ src/produit/clazz/ProduitNonToxique.java
@@ -1,0 +1,7 @@
+package produit.clazz;
+
+public class ProduitNonToxique extends Produit {
+    public ProduitNonToxique(String nom, int prix, String marque) {
+        super(nom, prix, marque);
+    }
+}
