+++ src/produit/interfaze/Toxique.java
@@ -1,0 +1,4 @@
+package produit.interfaze;
+
+public interface Toxique {
+}
